Enc common: separation of CDEF
