+++ Source/Lib/Encoder/Codec/aom_dsp_rtcd.c
@@ -87,31 +87,15 @@ void setup_rtcd_internal(CPU_FLAGS flags) {
     av1_wedge_sign_from_residuals = av1_wedge_sign_from_residuals_c;
     if (flags & HAS_AVX2) av1_wedge_sign_from_residuals = av1_wedge_sign_from_residuals_avx2;
 
-    eb_cdef_find_dir = eb_cdef_find_dir_c;
-    if (flags & HAS_AVX2) eb_cdef_find_dir = eb_cdef_find_dir_avx2;
-
-    eb_cdef_filter_block = eb_cdef_filter_block_c;
-    if (flags & HAS_AVX2) eb_cdef_filter_block = eb_cdef_filter_block_avx2;
     eb_compute_cdef_dist = compute_cdef_dist_c;
     if (flags & HAS_AVX2) eb_compute_cdef_dist = compute_cdef_dist_avx2;
     eb_compute_cdef_dist_8bit = compute_cdef_dist_8bit_c;
     if (flags & HAS_AVX2) eb_compute_cdef_dist_8bit = compute_cdef_dist_8bit_avx2;
-    eb_copy_rect8_8bit_to_16bit = eb_copy_rect8_8bit_to_16bit_c;
-    if (flags & HAS_AVX2) eb_copy_rect8_8bit_to_16bit = eb_copy_rect8_8bit_to_16bit_avx2;
 
     eb_av1_compute_stats = eb_av1_compute_stats_c;
     if (flags & HAS_AVX2) eb_av1_compute_stats = eb_av1_compute_stats_avx2;
     eb_av1_compute_stats_highbd = eb_av1_compute_stats_highbd_c;
     if (flags & HAS_AVX2) eb_av1_compute_stats_highbd = eb_av1_compute_stats_highbd_avx2;
-    eb_cdef_filter_block_8x8_16 =
-        eb_cdef_filter_block_8x8_16_avx2; // It has no c version, and is only called in parent avx2 function, so it's safe to initialize to avx2 version.
-#ifndef NON_AVX512_SUPPORT
-    if (flags & HAS_AVX512F) {
-        eb_cdef_filter_block_8x8_16 = eb_cdef_filter_block_8x8_16_avx512;
-        eb_av1_compute_stats        = eb_av1_compute_stats_avx512;
-        eb_av1_compute_stats_highbd = eb_av1_compute_stats_highbd_avx512;
-    }
-#endif
 
     eb_av1_lowbd_pixel_proj_error  = eb_av1_lowbd_pixel_proj_error_c;
     eb_av1_highbd_pixel_proj_error = eb_av1_highbd_pixel_proj_error_c;
